coordinator: Use a new SingletonException class for singleton exceptions
The singleton exceptions in Coordination contained meaningless stack
traces upon construction and were also mutable: a cause could be set and
the stack trace filled in or set. A new package private SingletonException
class is defined that is immutable (except for suppressed exceptions
introduced in Java 7). This exception class is then used for the
singleton exceptions in Coordination.

The package version is bumped to 1.0.2 since this did not change the
public API of the package but did change the implementation.